feat: limit AbsoluteCount of threshold to 100

+++ contracts/dao-members/src/msg.rs
@@ -7,7 +7,7 @@ use cw_utils::{Duration, Threshold};
 pub struct InstantiateMsg {
     pub members: Vec<Member>,
     pub dao_name: String,
-    pub threshold_percentage: Decimal,
+    pub threshold_percentage: u64,
     pub max_voting_period: Duration,
     pub governance_addr: Addr,
 }
